[qt] Merge ClientManagerSetup/Events back into ClientManager
testConnection, signup, subscribeToEvent, and unsubscribeFromEvent
were isolated in separate TUs solely to avoid MSVC C1202 from rfl
template instantiation depth.  With a TU-local #pragma constexpr_steps
(1 000 000) now in place, the split is unnecessary.  Merge the four
methods back and delete the two workaround files.
